+++ packages/definitions/src/constants.ts
@@ -28,3 +28,4 @@ export const ANCHOR_HINTS_DATA_ATTR_WHITELIST = [
 export const EVENT_OPEN_GROUP_MANAGER = "eye-note-open-group-manager";
 export const EVENT_OPEN_QUICK_MENU = "eye-note-open-quick-menu";
 export const EVENT_OPEN_SETTINGS_DIALOG = "eye-note-open-settings-dialog";
+export const EVENT_OPEN_NOTIFICATIONS_PANEL = "eye-note-open-notifications-panel";